Mp: Added kotlinx serialization and renamed data package to model
