+++ android/src/main/java/com/akjaw/fullerstack/screens/list/NotesListController.kt
@@ -4,12 +4,12 @@ import base.usecase.Either
 import base.usecase.UseCaseAsync
 import com.akjaw.fullerstack.screens.common.navigation.ScreenNavigator
 import com.akjaw.fullerstack.screens.common.toParcelable
-import data.Note
 import feature.noteslist.FetchNotes
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
+import model.Note
 
 class NotesListController(
     private val screenNavigator: ScreenNavigator,
