+++ android/src/test/java/com/akjaw/fullerstack/screens/editor/NoteEditorControllerTest.kt
@@ -2,7 +2,6 @@ package com.akjaw.fullerstack.screens.editor
 
 import com.akjaw.fullerstack.screens.common.navigation.ScreenNavigator
 import com.akjaw.fullerstack.screens.common.toParcelable
-import data.Note
 import feature.noteslist.AddNote
 import feature.noteslist.UpdateNote
 import helpers.validation.NoteInputValidator
@@ -13,6 +12,7 @@ import io.mockk.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestCoroutineDispatcher
 import kotlinx.coroutines.test.TestCoroutineScope
+import model.Note
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
