+++ shared/src/commonMain/kotlin/model/Note.kt
@@ -1,7 +1,7 @@
-package data
+package model
 
 import com.soywiz.klock.DateTime
-// TODO rename package to model
+
 data class Note(
     val id: Int = -1,
     val title: String = "",
