+++ shared/src/commonMain/kotlin/network/NoteApi.kt
@@ -1,6 +1,6 @@
 package network
 
-import data.Note
+import model.Note
 
 interface NoteApi {
     suspend fun getNotes(): List<Note>
